requirements.txt
Django==2.2.6
djangorestframework==3.10.3
djangorestframework-bulk==0.2.1
djoser==2.0.3
django-cors-headers==3.1.1
django-filter==2.2.0
easy-thumbnails==2.7
transliterate==1.10.2
django-simple-history==2.10.0
beautifulsoup4==4.9.3
requests==2.25.1
selenium==3.141.0
